Hard cap the mothership's cumulative FOV on level-up.
Long-range targeting may need to work differently for motherships.  Before the hard cap the motherships would accumulate a lot of additional points to their FOV when other motherships were destroyed. This was to keep the game moving (motherships seeking out targets instead of just sitting around).

We can use the `Arena.swarms` dict to select long-range targets when other targets are not in view. Since `Arena.swarms` is always a relatively small dict with a consistent naming pattern it should be a simple patch.